integrate new FRI
