+++ prover/prover/benches/binary_merkle_tree.rs
@@ -3,29 +3,26 @@
 use std::iter::repeat_with;
 
 use binius_field::Random;
-use binius_prover::{
-	hash::parallel_compression::ParallelCompressionAdaptor,
-	merkle_tree::{MerkleTreeProver, prover::BinaryMerkleTreeProver},
-};
+use binius_prover::merkle_tree::MerkleTreeProver;
+use binius_transcript::ProverTranscript;
 use binius_verifier::{
-	config::B128,
-	hash::{StdCompression, StdDigest},
+	config::{B128, StdChallenger},
+	hash::{PseudoCompressionFunction, StdCompression, StdDigest},
 };
 use criterion::{Criterion, Throughput, criterion_group, criterion_main};
-use digest::{FixedOutputReset, Output, core_api::BlockSizeUser};
+use digest::{Digest, Output, core_api::BlockSizeUser};
 
 const LOG_ELEMS: usize = 17;
 const LOG_ELEMS_IN_LEAF: usize = 4;
+const COMMIT_LAYER: usize = 0;
 
 type F = B128;
 
 fn bench_binary_merkle_tree<H, C>(c: &mut Criterion, compression: C, hash_name: &str)
 where
-	H: binius_prover::hash::ParallelDigest<Digest: BlockSizeUser + FixedOutputReset>,
-	C: binius_verifier::hash::PseudoCompressionFunction<Output<H::Digest>, 2> + Sync,
+	H: Digest + BlockSizeUser,
+	C: PseudoCompressionFunction<Output<H>, 2> + Sync,
 {
-	let parallel_compression = ParallelCompressionAdaptor::new(compression);
-	let merkle_prover = BinaryMerkleTreeProver::<_, H, _>::new(parallel_compression);
 	let mut rng = rand::rng();
 	let data = repeat_with(|| F::random(&mut rng))
 		.take(1 << (LOG_ELEMS + LOG_ELEMS_IN_LEAF))
@@ -36,9 +33,21 @@ where
 	));
 	group.sample_size(10);
 	group.bench_function(
-		format!("{} log elems size {}xB64 leaf", LOG_ELEMS, 1 << LOG_ELEMS_IN_LEAF),
+		format!("{} log elems size {}xB128 leaf", LOG_ELEMS, 1 << LOG_ELEMS_IN_LEAF),
 		|b| {
-			b.iter(|| merkle_prover.commit(&data, 1 << LOG_ELEMS_IN_LEAF));
+			// create prover transcipt
+			let challenger = StdChallenger::default();
+			let mut prover_transcript = ProverTranscript::new(challenger);
+
+			b.iter(|| {
+				let _merkle_prover = MerkleTreeProver::<_, H, _>::write_commitment(
+					compression.clone(),
+					data.clone(), // NOTE: The benchmark also measures the cloning.
+					LOG_ELEMS_IN_LEAF,
+					COMMIT_LAYER,
+					&mut prover_transcript.message(),
+				);
+			});
 		},
 	);
 	group.finish()
