+++ prover/prover/src/error.rs
@@ -1,9 +1,6 @@
 // Copyright 2025 Irreducible Inc.
 
-use crate::{
-	fri,
-	protocols::{intmul, shift, sumcheck},
-};
+use crate::protocols::{intmul, shift, sumcheck};
 
 #[derive(Debug, thiserror::Error)]
 pub enum Error {
@@ -11,8 +8,6 @@ pub enum Error {
 	ArgumentError { arg: String, msg: String },
 	#[error("sumcheck error: {0}")]
 	Sumcheck(#[from] sumcheck::Error),
-	#[error("FRI error: {0}")]
-	Fri(#[from] fri::Error),
 	#[error("transcript error: {0}")]
 	Transcript(#[from] binius_transcript::Error),
 	#[error("integer multiplication error: {0}")]
