+++ verifier/verifier/src/lib.rs
@@ -3,10 +3,8 @@ pub mod and_reduction;
 pub mod config;
 mod error;
 pub mod fri;
-pub mod fri2;
 pub mod hash;
 pub mod merkle_tree;
-pub mod merkle_tree2;
 pub mod pcs;
 pub mod protocols;
 pub mod ring_switch;
